requirements.txt
fastapi
uvicorn
pandas
# libomp-dev
faiss-cpu
sentence-transformers
# faiss-gpu
# unicodedata
transformers
torch
openpyxl
